BackCheck/requirements.txt
Flask~=3.0.3
Flask-Cors==4.0.0
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv~=1.0.0
joserfc~=0.9.0
supabase~=2.3.4
app~=0.0.1
gotrue~=2.1.0
bycypt~=4.1.2